model-api/requirements.txt
requests
torch
torchvision
ultralytics
fastapi
uvicorn
Pillow
requests
matplotlib
numpy
python-multipart
